Add files via upload
Added the Blueprint search tab. this now lets you pulls all available blueprints for the selected user. I have included optional selections for time and material efficacy. Please note they are not to be taken as fact(station and other factors may alter the final number of materials needed or time to produce. still pending the character blueprints command to get updated and actually allow us to see the final input materials and time based on skill and research.